+++ GreensFunction1DRadAbs.cpp
@@ -127,7 +127,7 @@ uint GreensFunction1DRadAbs::guess_maxi(Real const& t) const
 
     if (thr <= 0.0)
     {
-        return MAX_TERMS;
+        return static_cast<uint>(MAX_TERMS);
     }
 
     const Real max_root( sqrt(root0 * root0 - log(thr) / Dt) );
@@ -135,10 +135,10 @@ uint GreensFunction1DRadAbs::guess_maxi(Real const& t) const
     const uint maxi(std::max( safety + 
                               static_cast<uint>
                               (max_root * L  / M_PI),
-                              MIN_TERMS )
+                              static_cast<uint>(MIN_TERMS) )
                     );
 
-    return std::min(maxi, MAX_TERMS);
+    return std::min(maxi, static_cast<uint>(MAX_TERMS) );
 }
 
 
@@ -255,7 +255,7 @@ Real GreensFunction1DRadAbs::p_survival_table(Real t, RealVector& psurvTable) co
 
     const uint maxi( guess_maxi(t) );
     
-    if( maxi >= MAX_TERMS )
+    if( maxi >= static_cast<uint>(MAX_TERMS) )
         log_.warn("drawT: maxi was cut to MAX_TERMS for t = %.16g", t);
     
     if ( psurvTable.size() < maxi )
@@ -409,7 +409,7 @@ Real GreensFunction1DRadAbs::prob_r (Real r, Real t) const
     uint n = 0;
     do
     {
-        if ( n >= MAX_TERMS )
+        if ( n >= static_cast<uint>(MAX_TERMS) )
         {
             log_.warn("Too many terms needed for prob_r. N: %5u", n);
             break;
@@ -426,7 +426,7 @@ Real GreensFunction1DRadAbs::prob_r (Real r, Real t) const
     }
     while (fabs(term/sum) > EPSILON*PDENS_TYPICAL || 
            fabs(prev_term/sum) > EPSILON*PDENS_TYPICAL ||
-           n < MIN_TERMS );
+           n < static_cast<uint>(MIN_TERMS) );
 
     return 2.0*exp(vexpo)*sum;
 }
@@ -460,7 +460,7 @@ Real GreensFunction1DRadAbs::flux_tot (Real t) const
     uint n = 0;
     do
     {
-        if ( n >= MAX_TERMS )
+        if ( n >= static_cast<uint>(MAX_TERMS) )
         {
             log_.warn("Too many terms needed for flux_tot. N: %5u", n );
             break;
@@ -474,7 +474,7 @@ Real GreensFunction1DRadAbs::flux_tot (Real t) const
     }
     while (fabs(term/sum) > EPSILON*PDENS_TYPICAL ||
            fabs(prev_term/sum) > EPSILON*PDENS_TYPICAL ||
-           n < MIN_TERMS );
+           n < static_cast<uint>(MIN_TERMS) );
 
     return 2.0*D*exp(vexpo)*sum;
 }
@@ -729,7 +729,7 @@ Real GreensFunction1DRadAbs::p_int_r_table(Real const& r,
     const Real vexpo(-v*v*t/4.0/D - v*r0/2.0/D);
     const Real prefac( 2.0*exp(vexpo) );
 
-    if( maxi >= MAX_TERMS )
+    if( maxi >= static_cast<uint>(MAX_TERMS) )
     {
         log_.warn("drawR: maxi was cut to MAX_TERMS for t = %.16g", t);
         std::cerr << dump();
@@ -744,7 +744,7 @@ Real GreensFunction1DRadAbs::p_int_r_table(Real const& r,
 
     p = funcSum(boost::bind(&GreensFunction1DRadAbs::p_int_r_i, 
                             this, _1, r, t, table),
-                MAX_TERMS);
+                static_cast<uint>(MAX_TERMS) );
 
     return prefac * p;
 }
